Added name and student id to authors.yaml.
Current edits are geared towards making a MCTS implementation.

Updated student_agent.py to contain a class Node representing a node in the SearchTree.

A Node will contain the state it represents, the playout wins/totalPlays, the children of that node (representing states that can be transitioned to),
which player’s turn it is, and the parent of that node.

The Node class has a method to compute all the children of a node.
It creates a Node for every possible state that can be transitioned to from the
current state for the player whos turn it is.